Cast int to pointer properly

+++ Pin.cpp
@@ -6,7 +6,7 @@ Pin::Pin(uint8_t number): number(number)
 {
 	port			= pgm_read_byte(digitalPinToPortPGM + (number));
 	bitMask			= pgm_read_byte(digitalPinToBitMaskPGM + (number));
-	timer			= (volatile uint16_t *)(pgm_read_byte(digitalPinToTimerPGM + number));
+	timer			= (volatile uint16_t *)(pgm_read_byte(digitalPinToTimerPGM + (intptr_t)number));
 	analogBit		= number;
 	modeRegister	= (volatile uint16_t *)(pgm_read_word(portToModePGM + port));
 	inputRegister	= (volatile uint16_t *)(pgm_read_word(portToInputPGM + port));
